+++ packages/expo-dev-menu/vendored/react-native-safe-area-context/ios/SafeAreaView/DevMenuRNCSafeAreaViewMode.m
@@ -1,0 +1,11 @@
+#import <React/RCTConvert.h>
+#import "DevMenuRNCSafeAreaViewMode.h"
+
+@implementation RCTConvert (DevMenuRNCSafeAreaView)
+
+RCT_MULTI_ENUM_CONVERTER(DevMenuRNCSafeAreaViewMode, (@{
+  @"padding": @(DevMenuRNCSafeAreaViewModePadding),
+  @"margin": @(DevMenuRNCSafeAreaViewModeMargin),
+}), DevMenuRNCSafeAreaViewModePadding, integerValue);
+
+@end