+++ packages/expo-dev-menu/vendored/react-native-safe-area-context/src/index.tsx
@@ -1,0 +1,4 @@
+export * from './SafeAreaContext';
+export * from './SafeAreaView';
+export * from './InitialWindow';
+export * from './SafeArea.types';